10:59 26sept

+++ app-order-service/src/main/java/uz/pdp/apporderservice/payload/Aksverka.java
@@ -15,7 +15,7 @@ public class Aksverka implements Comparable<Aksverka>{
 
     private String productName;
 
-    private Integer productCount;
+    private Double productCount;
 
     private Double productPrice;
